requirements.txt
eel
pillow
tflite-runtime
opencv-python
numpy<2